refactor: Move utils to fixtures
